Move methods to oslogin_utils, skip test on Windows images, generalize getent tests to not be project-dependent
