+++ imagetest/fixtures.go
@@ -379,7 +379,11 @@ func (t *TestVM) AddMetadata(key, value string) {
 
 // AddScope adds the specified auth scope to the service account on the VM.
 func (t *TestVM) AddScope(scope string) {
-	t.instance.Scopes = append(t.instance.Scopes, scope)
+	if t.instance != nil {
+		t.instance.Scopes = append(t.instance.Scopes, scope)
+	} else if t.instancebeta != nil {
+		t.instancebeta.Scopes = append(t.instancebeta.Scopes, scope)
+	}
 }
 
 // RunTests runs only the named tests on the testVM.
