+++ imagetest/test_suites/oslogin/ssh_test.go
@@ -6,7 +6,6 @@ package oslogin
 import (
 	"context"
 	"fmt"
-	"os/exec"
 	"strings"
 	"testing"
 	"time"
@@ -14,17 +13,11 @@ import (
 	compute "cloud.google.com/go/compute/apiv1"
 	computepb "cloud.google.com/go/compute/apiv1/computepb"
 	secretmanager "cloud.google.com/go/secretmanager/apiv1"
-	secretmanagerpb "cloud.google.com/go/secretmanager/apiv1/secretmanagerpb"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/utils"
 	"golang.org/x/crypto/ssh"
 )
 
 const (
-	instanceMetadataURL       = "http://metadata.google.internal/computeMetadata/v1/instance"
-	osloginMetadataURL        = "http://metadata.google.internal/computeMetadata/v1/oslogin"
-	serviceAccountMetadataURL = "service-accounts/default/email"
-	projectZoneURL            = "zone"
-
 	// users
 	nopermsUser   = "non-user"
 	normalUser    = "normal-user"
@@ -43,75 +36,16 @@ const (
 	admin2FASshKey   = "admin-2fa-ssh-key"
 )
 
-// Creates the posix name from the given username.
-func getPosix(user string) string {
-	return strings.Join(strings.FieldsFunc(user, getPosixSplit), "_")
-}
-
-func getPosixSplit(r rune) bool {
-	return r == '.' || r == '@' || r == '-'
-}
-
-// Gets the given secret.
-func getSecret(ctx context.Context, client *secretmanager.Client, secretName string) (string, error) {
-	// Get project
-	project, _, err := getProjectZone()
-	if err != nil {
-		return "", fmt.Errorf("failed to get project: %v", err)
-	}
-
-	// Make request call to Secret Manager.
-	req := &secretmanagerpb.AccessSecretVersionRequest{
-		Name: fmt.Sprintf("projects/%s/secrets/%s/versions/latest", project, secretName),
-	}
-	resp, err := client.AccessSecretVersion(ctx, req)
-	if err != nil {
-		return "", fmt.Errorf("failed to get secret: %v", err)
-	}
-	return string(resp.Payload.Data), nil
-}
-
-// Gets the project and zone of the instance.
-func getProjectZone() (string, string, error) {
-	projectZone, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("curl -H \"Metadata-Flavor: Google\" %s/%s", instanceMetadataURL, projectZoneURL)).Output()
-	if err != nil {
-		return "", "", err
-	}
-	projectZoneSplit := strings.Split(string(projectZone), "/")
-	project := projectZoneSplit[1]
-	zone := projectZoneSplit[3]
-	return project, zone, nil
-}
-
-// Gets the service account currently operating on the instance.
-func getServiceAccount() (string, error) {
-	serviceAccountBytes, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("curl -H \"Metadata-Flavor: Google\" %s/%s", instanceMetadataURL, serviceAccountMetadataURL)).Output()
-	if err != nil {
-		return "", fmt.Errorf("failed to get service account: %v", err)
-	}
-	serviceAccount := string(serviceAccountBytes)
-	return serviceAccount, nil
-}
-
-// Gets the name of the instance running the test.
-func getInstanceName() (string, error) {
-	name, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("curl -H \"Metadata-Flavor: Google\" %s/name", instanceMetadataURL)).Output()
-	if err != nil {
-		return "", err
-	}
-	return string(name), nil
-}
-
 // Changes the given metadata key to have the given value on the given instance.. If the key does not exist,
 // then this will create the key-value pair in the instance's metadata.
 func changeMetadata(ctx context.Context, client *compute.InstancesClient, key, value string) error {
-	vmname, err := getInstanceName()
+	vmname, err := getInstanceName(ctx)
 	if err != nil {
 		return fmt.Errorf("error getting vm name: %v", err)
 	}
 
 	// Get project and zone of instance.
-	project, zone, err := getProjectZone()
+	project, zone, err := getProjectZone(ctx)
 	if err != nil {
 		return err
 	}
